Rearrange/reroute skeleton to get settings menu links working.
