+++ src/components/Dashboard/index.jsx
@@ -42,7 +42,7 @@ const useStyles = makeStyles(theme => ({
 
 export default function Dashboard(props) {
   const classes = useStyles();
-  const { children, sidenav } = props;
+  const { children } = props;
 
   return (
     <Fragment>
@@ -73,7 +73,6 @@ export default function Dashboard(props) {
         </Toolbar>
       </AppBar>
       <div>Dashboard</div>
-      <div className={classes.sidenav}>{sidenav}</div>
       <main className={classes.main}>{children}</main>
     </Fragment>
   );
