--- src/components/SettingsNav/navItems.js
@@ -1,17 +1,0 @@
-export default [
-  {
-    value: 'Users',
-    path: '/settings/users',
-    id: 'users',
-  },
-  {
-    value: 'Roles',
-    path: '/settings/roles',
-    id: 'roles',
-  },
-  {
-    value: 'Required Signoffs',
-    path: '/settings/required_signoffs',
-    id: 'required_signoffs',
-  },
-];
