+++ src/views/Roles/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Dashboard from '../../components/Dashboard';
+
+export default function Roles() {
+  return (
+    <Dashboard>
+      <div>ROLES!</div>
+    </Dashboard>
+  );
+}
