--- src/views/Settings/RequiredSignoffs/index.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import Dashboard from '../../../components/Dashboard';
-import SettingsNav from '../../../components/SettingsNav';
-
-export default function RequiredSignoffs() {
-  return <Dashboard sidenav={<SettingsNav />}>REQUIRED SIGNOFFS!</Dashboard>;
-}
