--- src/views/Settings/index.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { Switch } from 'react-router-dom';
-import RouteWithProps from '../../components/RouteWithProps';
-import routes from './routes';
-
-export default function Settings(props) {
-  const {
-    match: { path },
-  } = props;
-
-  return (
-    <Switch>
-      {routes(path).map(({ routes, ...routeProps }) => (
-        <RouteWithProps key={routeProps.path || 'not-found'} {...routeProps} />
-      ))}
-    </Switch>
-  );
-}
