+++ src/views/Users/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Dashboard from '../../components/Dashboard';
+
+export default function Users() {
+  return (
+    <Dashboard>
+      <div>USERS!</div>
+    </Dashboard>
+  );
+}
